implements accepting order functionality

+++ src/app/brand/brands/brands.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { IBrand } from 'src/app/models/IBrand';
 import { BrandService } from './../brand.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-brands',
@@ -10,7 +11,7 @@ import { BrandService } from './../brand.service';
 export class BrandsComponent implements OnInit {
 
   brands:IBrand[];
-  constructor(private brandsService:BrandService) {
+  constructor(private brandsService:BrandService,private toasterService:ToastrService) {
     this.brands = [];
   }
 
@@ -26,9 +27,12 @@ export class BrandsComponent implements OnInit {
 
   onDelete(id:number)
   {
-    this.brandsService.deleteBrand(id).subscribe(data=>{
-      this.brands.splice(this.brands.findIndex(p => p.id === id), 1);    
-    });
+    if(window.confirm("Are you sure you want to delete this item ??")){
+      this.brandsService.deleteBrand(id).subscribe(data=>{
+        this.toasterService.success("Deleted succesfully");
+        this.brands.splice(this.brands.findIndex(p => p.id === id), 1);    
+      });
+    }
   }
 
 }
