+++ src/app/product/products/products.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { IProduct } from 'src/app/models/product';
 import { ShopParams } from 'src/app/models/ShopParams';
 import { ProductsService } from 'src/app/services/products.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-products',
@@ -14,7 +15,7 @@ export class ProductsComponent implements OnInit {
   totalCount: number;
   shopParams: ShopParams;
 
-  constructor(private productsService:ProductsService) { 
+  constructor(private productsService:ProductsService,private toasterService:ToastrService) { 
     this.products = [];
     this.totalCount = 0;
     this.shopParams = new ShopParams();
@@ -42,10 +43,12 @@ export class ProductsComponent implements OnInit {
   }
 
   deleteProduct(id: number) {
-    this.productsService.deleteProduct(id).subscribe((response: any) => {
-      this.products.splice(this.products.findIndex(p => p.id === id), 1);
-      this.totalCount--;
-    });
+    if(window.confirm("Are you sure you want to delete this item??")){
+      this.productsService.deleteProduct(id).subscribe((response: any) => {
+        this.toasterService.success("Item has been removed");
+        this.products.splice(this.products.findIndex(p => p.id === id), 1);
+        this.totalCount--;
+      });
+    }
   }
-
 }
